Set background color through Props and establish a default class for RedirectionButton component
