+++ src/components/Buttons/Buttons.css
@@ -1,5 +1,4 @@
-.button {
-  background-color: #c80502;
+.main-button {
   border-radius: 40px;
   padding: 11px 10px;
   color: #f7f7f7;
