--- src/components/RedirectionButton/RedirectionButton.tsx
@@ -1,24 +1,0 @@
-import { Link } from "react-router";
-import "./RedirectionButton.css";
-
-interface RedirectionButtonProps {
-  text: string;
-  className?: string;
-  redirection: string;
-  ariaLabel: string;
-}
-
-const RedirectionButton: React.FC<RedirectionButtonProps> = ({
-  text,
-  className,
-  redirection,
-  ariaLabel,
-}) => {
-  return (
-    <Link className={className} to={redirection} aria-label={ariaLabel}>
-      {text}
-    </Link>
-  );
-};
-
-export default RedirectionButton;
